Add files via upload
This program was my final project in course 1 of 5 while at the University of Washington Python certificate program.  It needs some further work, but I have left it as is, as a sort of bench-mark in time.  This was was the beginning of an exciting, new direction for me.  In subsequent edits of the program, I will redesign it into more of a bug tracking program that can serve as a business tool.